Move LDAP authenticator to separate package
