+++ presto-password-authenticators/src/main/java/io/prestosql/plugin/password/ldap/LdapAuthenticator.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.plugin.password;
+package io.prestosql.plugin.password.ldap;
 
 import com.google.common.base.VerifyException;
 import com.google.common.cache.CacheBuilder;
