+++ presto-password-authenticators/src/main/java/io/prestosql/plugin/password/ldap/LdapAuthenticatorFactory.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.plugin.password;
+package io.prestosql.plugin.password.ldap;
 
 import com.google.inject.Injector;
 import com.google.inject.Scopes;
