+++ presto-password-authenticators/src/main/java/io/prestosql/plugin/password/ldap/LdapConfig.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.plugin.password;
+package io.prestosql.plugin.password.ldap;
 
 import io.airlift.configuration.Config;
 import io.airlift.configuration.ConfigDescription;
