+++ presto-password-authenticators/src/test/java/io/prestosql/plugin/password/ldap/TestLdapConfig.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.plugin.password;
+package io.prestosql.plugin.password.ldap;
 
 import com.google.common.collect.ImmutableMap;
 import io.airlift.units.Duration;
